16._Redis_Labs/pyproject.toml
[tool.poetry]
name = "16-redis-labs"
version = "0.1.0"
description = ""
authors = ["Anders Latif <andl@kea.dk>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
redis = "^4.3.4"
redis-om = "^0.0.27"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
